requirements.txt
mcp>=1.0.0
fastmcp>=2.0.0
httpx>=0.25.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
anyio>=4.0.0
requests>=2.31.0
flask>=2.3.0
pyjwt>=2.8.0
cryptography>=41.0.8
bcrypt>=4.0.0